fix: widen empty-gap span in OPTIMIZED range histogram against int overflow
chooseOptimizedBucketCount widened the overall span to double but still
subtracted the empty-gap bracket keys (spanEndKey - spanStartKey) as int.
For full-range Integer attributes that subtraction wraps negative, drives
recomputedStep below zero and silently collapses the OPTIMIZED grid to two
buckets. Widen both operands to double, mirroring the overall-span guard,
and add a regression test exercising the gap-widening branch across the
full [Integer.MIN_VALUE, Integer.MAX_VALUE] span.

Also clarify the overallCount contract docs (Histogram, HistogramContract,
CacheableHistogram, CacheableHistogramContract): the value is the
bucket-occurrence sum for point/price histograms and for range histograms
under the frequency-equalised behaviors (EQUALIZED / EQUALIZED_OPTIMIZED),
and the distinct-record count only for the overlap behaviors
(STANDARD / OPTIMIZED).

Ref: #1245

+++ evita_engine/src/main/java/io/evitadb/core/query/extraResult/translator/histogram/producer/RangeHistogramDataCruncher.java
@@ -333,10 +333,11 @@ public class RangeHistogramDataCruncher {
 		if (bucketCount - emptyColumns <= 2) {
 			return 2;
 		} else if (emptyColumns >= 2) {
-			// widen each operand to double before subtracting so the full int range does not overflow first
+			// widen each operand to double before subtracting so the full int range does not overflow first —
+			// this applies to both the overall span and the empty-gap span bracketed by spanStartKey/spanEndKey
 			final double span = (double) maxKey - (double) minKey;
 			final double optimalStep = span / bucketCount;
-			final double recomputedStep = optimalStep + (spanEndKey - spanStartKey) / 2.0;
+			final double recomputedStep = optimalStep + ((double) spanEndKey - (double) spanStartKey) / 2.0;
 			final int newBucketCount = (int) Math.floor(span / recomputedStep) + 2;
 			return Math.max(2, Math.min(newBucketCount, bucketCount));
 		} else {
